+++ src/main/java/com/soa/service/busi/za/UpdateYlZa.java
@@ -11,6 +11,7 @@ import com.soa.service.BaseService;
 import static com.soa.service.BaseService.getSession;
 import com.soa.util.SystemUtil;
 import com.soa.util.sqjw.SqjwUtil;
+import java.io.File;
 import java.io.IOException;
 import javax.annotation.Resource;
 import org.slf4j.Logger;
@@ -62,29 +63,38 @@ public class UpdateYlZa extends BaseService {
             String file1_path = null;
             String file2_path = null;
             String file3_path = null;
-            if (name != null) {
+            String line = File.separator;//通用文件分割符
+            if (name != null && file != null) {
                 file_path = sqjwUtil.upLoad(file, "za0001_file_path1", "za0001_addyl", name);
                 if (sqjwUtil.deleteFile(sqjwUtil.getRelPath(in.getStringValue("jyxkz")))) {//删除以前的文件
                     in.putStringValue("jyxkz", file_path.toString());//数据库中保存的路径
                 }
+            } else {
+                in.putStringValue("jyxkz", in.getStringValue("jyxkz").replace(SystemUtil.getSysConfig("za0001_server_root") , ""));//当不修改图片时默认加上根目录，还没找到具体原因，暂时这样暴力解决
             }
-            if (name1 != null) {
+            if (name1 != null && file1 != null) {
                 file1_path = sqjwUtil.upLoad(file1, "za0001_file_path1", "za0001_addyl", name1);
                 if (sqjwUtil.deleteFile(sqjwUtil.getRelPath(in.getStringValue("ajhgz")))) {//删除以前的文件
                     in.putStringValue("ajhgz", file1_path.toString());
                 }
+            } else {
+                in.putStringValue("ajhgz", in.getStringValue("ajhgz").replace(SystemUtil.getSysConfig("za0001_server_root") , ""));//当不修改图片时默认加上根目录，还没找到具体原因，暂时这样暴力解决
             }
-            if (name2 != null) {
+            if (name2 != null&&file2!=null) {
                 file2_path = sqjwUtil.upLoad(file2, "za0001_file_path1", "za0001_addyl", name2);
                 if (sqjwUtil.deleteFile(sqjwUtil.getRelPath(in.getStringValue("cspmt")))) {//删除以前的文件
                     in.putStringValue("cspmt", file2_path.toString());
                 }
+            }else {
+                in.putStringValue("cspmt", in.getStringValue("cspmt").replace(SystemUtil.getSysConfig("za0001_server_root") , ""));//当不修改图片时默认加上根目录，还没找到具体原因，暂时这样暴力解决
             }
-            if (name3 != null) {
+            if (name3 != null&&file3!=null) {
                 file3_path = sqjwUtil.upLoad(file3, "za0001_file_path1", "za0001_addyl", name3);
                 if (sqjwUtil.deleteFile(sqjwUtil.getRelPath(in.getStringValue("gsxkz")))) {//删除以前的文件
                     in.putStringValue("gsxkz", file3_path.toString());
                 }
+            }else {
+                in.putStringValue("gsxkz", in.getStringValue("gsxkz").replace(SystemUtil.getSysConfig("za0001_server_root") , ""));//当不修改图片时默认加上根目录，还没找到具体原因，暂时这样暴力解决
             }
             update("modify_yl_za", in);
         } catch (IOException ex) {
